Move `quote` to `prefect.utilities.annotations`
