+++ src/prefect/utilities/collections.py
@@ -12,7 +12,6 @@ from typing import (
     Callable,
     Dict,
     Generator,
-    Generic,
     Hashable,
     Iterable,
     Iterator,
@@ -27,6 +26,9 @@ from typing import (
 from unittest.mock import Mock
 
 import pydantic
+
+# Moved to `prefect.utilities.annotations` but preserved here for compatibility
+from prefect.utilities.annotations import Quote, quote  # noqa
 
 
 class AutoEnum(str, Enum):
@@ -201,33 +203,6 @@ def batched_iterable(iterable: Iterable[T], size: int) -> Iterator[Tuple[T, ...]
         if not batch:
             break
         yield batch
-
-
-class Quote(Generic[T]):
-    """
-    Simple wrapper to mark an expression as a different type so it will not be coerced
-    by Prefect. For example, if you want to return a state from a flow without having
-    the flow assume that state.
-    """
-
-    def __init__(self, data: T) -> None:
-        self.data = data
-
-    def unquote(self) -> T:
-        return self.data
-
-
-def quote(expr: T) -> Quote[T]:
-    """
-    Create a `Quote` object
-
-    Examples:
-        >>> from prefect.utilities.collections import quote
-        >>> x = quote(1)
-        >>> x.unquote()
-        1
-    """
-    return Quote(expr)
 
 
 def visit_collection(
